+++ src/bos/server/controllers/v2/healthz.py
@@ -28,7 +28,7 @@ from bos.common.utils import exc_type_msg
 from bos.server.models.healthz import Healthz
 from bos.server import redis_db_utils
 
-DB = redis_db_utils.get_wrapper(db='options')
+DB = redis_db_utils.OptionsDBWrapper()
 
 LOGGER = logging.getLogger(__name__)