+++ src/bos/server/redis_db_utils/options_dbwrapper.py
@@ -1,8 +1,7 @@
-#!/usr/bin/env python
 #
 # MIT License
 #
-# (C) Copyright 2021-2022, 2024-2025 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2025 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -22,17 +21,38 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
+"""
+OptionsDBWrapper class
+"""
 
+from bos.common.types.options import OptionsDict
 
-from .filters import (ActualBootStateIsSet,
-                      ActualStateAge,
-                      BootArtifactStatesMatch,
-                      BOSQuery,
-                      DesiredBootStateIsNone,
-                      DesiredBootStateIsOff,
-                      DesiredConfigurationIsNone,
-                      DesiredConfigurationSetInCFS,
-                      HSMState,
-                      LastActionIs,
-                      OR,
-                      TimeSinceLastAction)
+from .dbwrapper import DBWrapper
+from .defs import Databases
+
+# We store all options as json under this key so that the data format is
+# similar to other data stored in the database, and to make retrieval of all
+# options simpler
+OPTIONS_KEY = 'options'
+
+class OptionsDBWrapper(DBWrapper[OptionsDict]):
+    """
+    Options database wrapper
+    """
+
+    @property
+    def db_id(self) -> Databases:
+        return Databases.OPTIONS
+
+    @property
+    def options_exist(self) -> bool:
+        return OPTIONS_KEY in self
+
+    def get_options(self) -> OptionsDict | None:
+        return self.get(OPTIONS_KEY)
+
+    def put_options(self, data: OptionsDict) -> OptionsDict | None:
+        return self.put(OPTIONS_KEY, data)
+
+    def patch_options(self, data: OptionsDict) -> OptionsDict | None:
+        return self._patch(OPTIONS_KEY, data)