+++ src/bos/server/redis_db_utils/session_status_dbwrapper.py
@@ -1,8 +1,7 @@
-#!/usr/bin/env python
 #
 # MIT License
 #
-# (C) Copyright 2021-2022, 2024-2025 Hewlett Packard Enterprise Development LP
+# (C) Copyright 2025 Hewlett Packard Enterprise Development LP
 #
 # Permission is hereby granted, free of charge, to any person obtaining a
 # copy of this software and associated documentation files (the "Software"),
@@ -22,17 +21,20 @@
 # ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR
 # OTHER DEALINGS IN THE SOFTWARE.
 #
+"""
+SessionStatusDBWrapper class
+"""
 
+from bos.common.types.session_extended_status import SessionExtendedStatus
 
-from .filters import (ActualBootStateIsSet,
-                      ActualStateAge,
-                      BootArtifactStatesMatch,
-                      BOSQuery,
-                      DesiredBootStateIsNone,
-                      DesiredBootStateIsOff,
-                      DesiredConfigurationIsNone,
-                      DesiredConfigurationSetInCFS,
-                      HSMState,
-                      LastActionIs,
-                      OR,
-                      TimeSinceLastAction)
+from .defs import Databases
+from .tenant_aware_dbwrapper import TenantAwareDBWrapper
+
+class SessionStatusDBWrapper(TenantAwareDBWrapper[SessionExtendedStatus]):
+    """
+    Wrapper for session status database
+    """
+
+    @property
+    def db_id(self) -> Databases:
+        return Databases.SESSION_STATUS